moving email functions to /lib
